split AboutSection to HeroSection, StatsSection, IntroSection
